Added foundational redux for clothing/pieces
Very rudimentary stuff, will probably have to change implementation of the slothesSlice in the future. Also need to look into efficienly making the piece list data without having to reverse and all that. The scroll is also not staying down, might be because all children have the same id, but could also be a bug to fix