Fixed scroll rubber-banding issue, better efficiency with loading clothinglist data
The rubber banding issue ended up being due to the absolute position, which is not needed in the first place. Also styled with some width and height of pieces relative to the width, as to be more consistent corss-device. Handling the clothinglist data is much better now, just add button to start and then everything else after. This is a good, simple, efficient solution that should lend well when we start sorting the lists in the future.