Implemented getting the piece by id in PieceScreen, a bit inefficient though
The PieceScreen currently has to get all pieces, unable to just select its info by id. Might want to make it a dictionary of IDs and piecetypes where you can just select instead of filtering through an array every time. That  but it should figure itself out. Maybe an ordered dictionary or something, or store an array and dictionary at the same time for the dual purposes.